Starting to split large module into sub-modules
