+++ pronto/views/api/signatures/__init__.py
@@ -1,6 +1,7 @@
 from flask import jsonify, request
 
 from pronto import app, db, xref
+from . import taxonomy
 
 
 def build_method2protein_query(signatures: dict, **kwargs):
@@ -504,109 +505,6 @@ def _prot_key(p):
 
 def _sign_key(s):
     return 0 if s["integrated"] else 1, s["accession"]
-
-
-@app.route("/api/signatures/<path:accessions_str>/taxonomy/")
-def get_taxonomic_origins(accessions_str):
-    accessions = []
-    for acc in accessions_str.split("/"):
-        acc = acc.strip()
-        if acc and acc not in accessions:
-            accessions.append(acc)
-
-    taxon_name = None
-    try:
-        taxon_id = int(request.args["taxon"])
-    except (KeyError, ValueError):
-        taxon_id = None
-    else:
-        cur = db.get_oracle().cursor()
-        cur.execute(
-            """
-            SELECT FULL_NAME
-            FROM {}.ETAXI
-            WHERE TAX_ID = :1
-            """.format(app.config["DB_SCHEMA"]),
-            (taxon_id,)
-        )
-        row = cur.fetchone()
-        cur.close()
-        if row:
-            taxon_name = row[0]
-        else:
-            return jsonify({}), 400  # TODO: return error
-
-    ranks = (
-        "superkingdom",
-        "kingdom",
-        "phylum",
-        "class",
-        "order",
-        "family",
-        "genus",
-        "species"
-    )
-    i = 0
-    try:
-        i = ranks.index(request.args.get("rank"))
-    except ValueError:
-        pass
-    finally:
-        rank = ranks[i]
-
-    query = """
-            SELECT E.TAX_ID, E.FULL_NAME, MT.METHOD_AC, MT.PROTEIN_COUNT
-            FROM {0}.METHOD_TAXA MT
-            LEFT OUTER JOIN {0}.ETAXI E 
-              ON MT.TAX_ID = E.TAX_ID
-        """.format(app.config["DB_SCHEMA"])
-
-    params = {"rank": rank}
-    for i, acc in enumerate(accessions):
-        params["acc" + str(i)] = acc
-
-    if taxon_id is not None:
-        query += """
-            INNER JOIN {}.LINEAGE L
-            ON E.LEFT_NUMBER = L.LEFT_NUMBER 
-            AND L.TAX_ID = :taxid
-        """.format(app.config["DB_SCHEMA"])
-        params["taxid"] = taxon_id
-
-    query += """
-        WHERE MT.METHOD_AC IN ({}) 
-        AND MT.RANK = :rank
-    """.format(
-        ','.join([":acc" + str(i) for i in range(len(accessions))]),
-    )
-
-    cur = db.get_oracle().cursor()
-    cur.execute(query, params)
-    taxa = {}
-    for tax_id, tax_name, acc, n_proteins in cur:
-        if tax_id in taxa:
-            taxa[tax_id]["signatures"][acc] = n_proteins
-        else:
-            taxa[tax_id] = {
-                "id": tax_id,
-                "name": tax_name if tax_id else "Others",
-                "signatures": {acc: n_proteins}
-            }
-
-    cur.close()
-
-    return jsonify({
-        "taxa": sorted(taxa.values(), key=_taxa_key),
-        "rank": rank,
-        "taxon": {
-            "id": taxon_id,
-            "name": taxon_name
-        }
-    })
-
-
-def _taxa_key(t):
-    return 1 if t["id"] is None else 0, -sum(t["signatures"].values())
 
 
 @app.route("/api/signatures/<path:accessions_str>/descriptions/")
